feat!(api): `BaseCalculator` class for calculations
**Problem**

In the future, Piquasso will support calculations using Tensorflow.

In order to achieve this without copying huge amount of existing code,
some smaller calculations are needed to be injected into the
simulations.

Among these calculations are the functions for the permanent, hafnian
and loop hafnian calculations, which are also present in `Config`.
However, the small calculations to be injected are not practical to
store in `Config`.

**Solution**

The calculations for permanent, hafnian and loop hafnian were moved from
`Config` to a separate `BaseCalculator` class, which is injected into
`Simulator`, `State`, `FockSpace` and `DensityMatrixCalculation`
classes.

The `BaseCalculator` is not customizable, and its subclasses are not
made available through the public API.

**BREAKING CHANGE**

By enabling Tensorflow support in the future, we dropped support for
customizing the permanent, hafnian and loop hafnian calculations through
`Config`.

Reasons:
1. The user could break Tensorflow calculations by customized
calculations;
2. The resulting code would be harder to maintain: there would be
customizable and non-customizable calculations as well;
3. The API would get a bit more confusing. With this change the rule of
thumb is that simulation constants (e.g. Plack constant) would be stored
in `Config`, while calculations would be stored in `Calculator`.