perf(fock/pure): Economical gate application (#197)
**Problems**

1) Before this, the `_apply_subsystem_representation_to_state` has applied
the matrix on the subspace to the whole state via two `for` loops on the
whole Fock space, which is not really economical. In the second for
loop, there was an `if` condition whether or not the state vector should
be evolved.

2) Moreover, the indices of the state vectors and subspace matrices has
been calculated by looking up the index of a given basis vector in a
`FockSpace` instance.

**Solutions**

1) One could remove the `if` condition by being smart about the second
`for` loop: the second `for` loop should be iterating over the Fock
subspace instead of the whole Fock space.

2) A `get_index_in_fock_space` function has been written to calculate
explicitely the index instead of the lookup.

In this method, several variables got renamed to make more sense and
also the `_apply_subsystem_representation_to_state` has been renamed to
`_apply_subspace_matrix_to_state`. Moreover, several benchmarks have
been written.