refact(gates): Delete `_ScalableFockGates` (#196)
The class `_ScalableFockGates` got deleted, since this class should not
exist. The two subclasses `CubicPhase` and `Kerr` have different
parameters, and the current logic inserted e.g. a redundant `"gamma":
None` entry in the `Kerr._extra_params`.

Instead, the logic has been separated to the two subclasses `CubicPhase`
and `Kerr`.

Also, it is very confusing that there is two `"gamma"` and `"xi"` keys
in `CubicPhase` and `Kerr` respectively, precisely one in the `_params`
and one in the `_extra_params` attribute. The one in the `_extra_params`
attribute got a `_vector` suffix to avoid ambiguity.