FIX(decompositions): Williamson dtype fix
**Problem**

Sometimes scipy.linalg.sqrtm returns complex matrices even when real matrices with
positive eigenvalues got specified. This test specifies a matrix for which the
square root turns out to be complex even if it is real and positive definite.

When a complex matrix is specified to the schur algorithm, automatically the complex
schur algorithm is executed instead of the desired real schur algorithm.

**Solution**

The result is cast to real after the square root.
A test has been added for a specific matrix.