FIX(fock): Fidelity precision
**Problem**

In `scipy` version 1.10.0 `sqrtm` returned with an array with type
`complex256` but `numpy.linalg` returned with the following error:
```
TypeError: array type complex256 is unsupported in linalg
```
Casting it to `complex128` resulted in inaccurate fidelity values,
however.

**Solution**

A less computationally demanding expression for the fidelity is used.
For comparison, an expression for the fidelity between pure states have
been also implemented.