perf(tensorflow): Custom gradient for squeezing
**Problem**

The squeezing gradient calculation was slow because the computational
graph contained too much nodes. This could be reduced by defining a
custom gradient.
For testing Squeezing with phaseshifting gradient the `quadratures_mean_variance` needed to be compatible with Tensorflow as well.
Additionally a possible bug in Tensorflow arose regarding the `tf.Variable` `dtype`

**Partial solution**

The custom gradient for the single mode squeezing matrix got defined.
For the phaseshift gradient we must take its conjugate when multiplying with the upstream due to Tensorflow implementations.
For `quadratures_mean_variance` the solution was to rewrite the density matrix to a list form.
The `dtype` error is currently resolved by casting the gradient to `float32` at the end of `create_single_mode_squeezing_gradient`

**Sources**

- https://quantum-journal.org/papers/q-2020-11-30-366/
- https://stackoverflow.com/questions/42479024/how-tensorflow-handles-complex-gradient