perf(purefock): Interferometer optimization (#218)
**Problem**

Creating the interferometer was very costly due to indexing, and the
algorithm calculating the matrix elements has been calculating these
indices redundantly. This algorithm is further compicated by the fact
that one has to search for the non-zero indices in a given occupation
number, in order to acquire certain indices.

**Solution**

The algorithm has been optimized: the `get_index_in_fock_subspace` call
for the column indices in the inner for loop is unnecessary, since it
has been done for the row indices in the outer for loop, and can be
reused. This way, the logic of the algorithm can be completely modified,
as follows:

1. All the index calculations are done in a separate for loop, and is
done only once for every block matrix corresponding to a given particle
number.

2. Knowing all the indices, using NumPy's advanced indexing and einsum,
one can write a more efficient algorithm, calculating a single matrix
column in every iteration.