feat(script): Script for norm value monitoring
**Problem**
As we stack the amount of layers in CVNNs the possibility of losing some
information during calculation increases too. We therefore should
profile how this changes compared to Strawberry Fields.

**Solution**
We wrote a new script containing both a Piquasso and a Strawberry Fields
benchmark with increasing number of layers, modes, and cutoff values.
For this, the `PureFockState` class got slightly extended, so that each
time the state gets normalized it saves its current norm into an
internal list, which is used during the benchmark. For Strawberry Fields
an exact formula is used inside the gradient tape.