FIX(fock/pure): State vector dtype fix
After the `Vacuum` instruction, the initial state is inititalized without
`dtype` parameter, so it defaulted to `float64`. It haven't caused any errors,
because it only happened when a passive gate is applied to a vacuum, which
leaves the vacuum invariant, and for active gates it is cast to complex. Still,
it is technically incorrect.