Generating the Hemiltonian
**Problem**

A possible group of gates for gate synthesis is which consists of gates
built upon polynomials of position (x) and momentum (p) operators. These
operators do not commute in the regular way, so constructing them in a
way where they form a self-adjoint matrix is not straight-forward.

**Solution**
For this, I based the generation on a binary tree-like ordering of
different permutations of x and p operators for each term and each
degree of polynomial. Currently, the coefficient generation is very
basic, it generates random numbers, and then sets the corresponding pair
of coeffient to it's conjugate, or simply generates a real number for
symmetric permuations. For greater detail see the comment on top of
`hemiltonian_generation.py`.

**Future Work**
Read the coefficients from a file, and check if they generate a
self-adjoint matrix or not. Additionally, the generation needs
improvements, alongside the mathematical foundation, since the current
question is whether this conjugate-pair coefficient approach covers
every Hamiltonian or not. Put differently: Is every H Hamiltonian in its
polynomial form is such that the symmetric terms can only have real
values, and that each other term has a conjugate counterpart? If so,
then this generative method in theory could cover any H.