Hamiltonian with different operators
** Problem **

Generating the unitary matrix for the gate sytnthesis problem based on
position and momentum operators proved to be slightly convoluted due to
their commutational relations. Additionaly the whole prototype solution
in the previous commit (a703fc8fb523ac2642b0cbf066c2865035b17ff4) was
overcomplicated in a variety of terms like parametrization and
redundancy in general.

** Solution **

As a second attempt this time we use only the creation and annihilation
operators which have a more suitable and managable commutational relation.
Here we use normal ordering for the polynomial, meaning in all terms
every creation operator is "before" the annihilation operator.
Additionally coefficients are generated in the following orderly
fashion:
`ad^k + ad^(k-1)*a + ... + ad*a^(k-1) + a^k + c`
where `ad`,`a`,`k`, and `c` are the creation operator, annihilation
operator, degree of polynomial, and the constant term correspondingly