perf(tensorflow): Converting upstream to numpy
The upstream differential in a custom gradient does not always have a value,
but when it has, one can calculate using numpy.

The function `get_static_value` evaluates the `tensorflow` tensor if it is
efficiently calculable, and we proceed with the calculation using `numpy`
arrays. If it is not efficiently calculable (returns with `None`), we calculate
using `tensorflow`.