perf(tensorflow): Interferometer gradient
Some indexing has been moved outside of the for loop to improve performance.

Moreover, if the upstream is static valued, then we calculate with numpy,
meaning that we don't need to pad the matrices, since in numpy we can do item
assignments.

Finally, some caching has been added to some index calculation functions.