Reproducing Jupyter Notebook
Based on a gate-synthesis jupyter notebook from Strawberry Fields we
tried to reproduce the same convergence and methods in
`sf_gate_synthesis.py`.
As it turns out, our squeezing, and possibly displacement gradient
calculations are incorrect at the moment of this commit. However, a
temporary fix is provided in a different commit on this branch following
up this one. With that addition, we successfully reproduced the paper.
The following steps are aimed at using our own unitaries to generate
training data for the following classical neural network in the future.

Notebook source: https://strawberryfields.ai/photonics/demos/run_gate_synthesis.html#references