fix(gradient): Gradient conjugates fix
**Problem**
Certain conjugates during gradient calculations
were missing leading to incorrect gradient values.
Furthermore, there was a sign error in the squeezing
gradient.

**Solution**
These conjugates are now present, and squeezing
calculation is added. Additionally, the test in
which the problem surfaced got updated, and an extra
test is present for testing the conjugate validity
elsewhere.