chore(gate synthesis): Refactor
The whole gate synthesis pipeline got reworked into more and somewhat
differently shaped modules. The basic mechanism and ordering is the
same, but this is more flexible to changes, and easier to benchmark
different quantum gate implementations or vectorization attempts in the
future.

`gate_synthesis_main.py` is responsible for calling every other step in
the process, and every hyperparameter is set here as well.

`unitary_generation.py` has every method for generating the necessary
target variables for the CVNN. Coefficient generation now can be
modified easily by changing `_generate_all_random_coefficients` or the
subsequently called `_generate_random_coefficients_of_degree` functions.

`persistence.py` simply does all the data reading, and saving occuring
in the processes.

`cvnn_approximation.py` after receiving necessary data and parameters
begins approximating the target ket vectors inside `approximate_kets`.
Here, a `_perform_machine_learning` method is called for each unitary.
Inside the `GradientTape` one can change the second parameter (a function)
of the `_cost_function` in order to use a differently constructed CVNN.

`gates_and_gradients.py` contains the gates, and their gradients as
buildingblocks of the CVNN. Every aspect and approach can be organized
into a class of methods, this way, as long as the methods are called the
same, in the `cvnn_approcimation.py` switching between two variants of
gate construction can be done as easily as rewriting the right hand side
of a variable.