feat(cvnn): Decent running time version
This version contains many different solutions in order to decrease
computation time. This version proved to be fairly decent to make a
commit out of it.
Benchmarking got enhanced.
New `tf.function` and jit compilation functionalities are added.
A half complete vectorization attempt is present as well, but it turned
out to be less efficient at the moment.