backup(gate_synthesis): CVNN Approximation Data V1
Up until now the pipeline consisted of generating unitaries, and
corresponding coefficients, followed by performing machine learning on
every CVNN for each unitary for a given precision, and then saving the
gate parameters as a result to train the classical neural network. This
commit serves as a backup to save these results, since now we pivot into
a slightly different approach.