feat(fock): Avoid matrix embedding
In `FockSimulator`, the unitary matrices acting on the density operator were
embedded in the Fock space, before directly multiplying with the density
matrix. This can be done faster, by avoiding embedding and modifying only the
relevant parts of the density matrix, as done in `PureFockSimulator` with the
state vector instead of the density matrix. This approach has been implemented
in this patch.