FIX(tensorflow): Passive linear gradient error
Simply put, when multiplying a vector with a fix matrix as `M @ v`, its
jacobian by `v` is just `M`, but in the forward pass, we have to return `M.T @
upstream` in the custom gradient, but `M @ upstream` has been calculated (if
`M`, `v` would be real). The fix amounts to rewriting an einsum string. When
calculating the gradient of active transformations, this has been done
correctly.