feat!(clements): Rewrite Clements decomposition
The Clements decomposition in Piquasso is a bit hard to use in some cases. It
is not differentiable, and the datatypes in it are not trivial, it takes time
to wrap one's head around it.

For variational circuits, it would be beneficial to support calculating weights
from interferometer matrices and to make this decomposition differentiable.
These features have been enabled in this patch.