fix(gradients): Gradient floating point precision
The gradient of the squeezing gate by the amplitude of the squeezing parameter
has been cast to `float32`, but it is not needed anymore. It may have been
needed due to a type promotion we missed, but it seems that this issue has been
resolved now by `207d1f3b3aa28e0b13d1d85c3f5b5d2005f90245`.