feat(tensorflow): Choose NumPy API in forward pass
`tf.function` requires Piquasso to use Tensorflow's NumPy API in the forward
pass. By default, we use regular NumPy for performance reasons, but this patch
enables users to run Piquasso inside `tf.function`, which might be faster with
JIT compilation enabled. JIT compilation takes some time though, but the
compiled function is an order of magnitude faster than the original function
decorated with `tf.function` only.

In order to enable Piquasso to use `tf.function` the possibility of avoiding
the custom gradient calculations had to be implemented, which is implemented
through `calculator.no_custom_gradient`. In the calculations where the gradient
has been replaced by a custom one, a different numpy
(`calculator.forward_pass_np`) needs to be used, and its value depends on
`no_custom_gradient`.

Moreover, some functions might be slow for vague reasons, e.g., it is better to
avoid `tf.transpose` and `tf.gather(..., axis=1)`. These functions are avoided
where possible. Also, some functions work a bit differently, e.g.,
`tf.math.pow` handles `0^0` differently than `np.pow`, which is encountered in
`gate_matrices.py`.

Furthermore, the compilation time is in correlation with the graph size
compiled with Autograph. In order to limit the size of the compiled graph,
several things are implemented:

1. `decorate_with`

Decorating some functions in the simulation with `tf.function` during the
simulation makes the graph smaller, since consequent calls to this function
won't induce retracing. Therefore the possibility to decorate some of these
functions got enabled with the `decorate_with` constructor parameter of
`TensorflowCalculator`. In it, one can pass the `tf.function` decorator to
decorate the functions with, and one may also pass other parameter (like
`jit_compile`, `reduce_retracing`).

Since the compiled functions are cached within `TensorflowCalculator`, the
instance should be preserved between calculations. Moreover, a custom
`_TrivialTraceType` had to be implemented for `TensorflowCalculator` to make
AutoGraph avoid retracing.

Moreover, the `no_custom_gradient` needs to be switched on when a decorator is
provided.

2. `TensorArray`

`tf.function` does not handle Python for loops well, because each iteration in
the for loop is handled independently in Autograph. Therefore, one needs to
avoid using for loops as much as possible, and use vectorized computations. If
not possible, one can use `TensorArray` instead, which has been implemented in
`TensorflowCalculator.accumulator` and in `gate_matrices.py`.

Technically, it could be implemented in other places as well, but it is a bit
more challenging, and (hopefully) using `decorate_with` limits the graph size
well enough.