fix: Printing of instructions
Printing of some instructions yields an error, because `Instruction.__init__`
has not been called. This can be resolved by adding this call, or alternatively
by fixing `Instruction.__repr__`, which is a more forgiving solution.

Note, that deleting `Vacuum.__init__` for example is not a solution, since we
want to avoid the possibility for users to pass parameters directly in
`Instruction.__init__`.

Some tests have been written testing the string representation of some
instructions.