feat!(fock): Refactor `FockSpace`
The logic around `piquasso._math.fock` was quite a bit difficult.

1. The `FockSpace`, `FockBasis`, `FockOperatorBasis` classes overloaded builtin
types, which were, frankly, quite unnecessary.

2. The `FockSpace` object is responsible for storing the occupation numbers, in
an order which prescribes the ordering of the state vector or the density
matrix. However, since calculations depend on `BaseCalculator` and `Config`
objects, it was convenient to store these as attributes. This made copying this
object a bit involved.

3. Creating a `FockSpace` can be expensive, therefore a caching has been
implemented in `BaseFockState`, but this would be unnecessary if one would
refactor `FockSpace` into a function and decorate it with `lru_cache`.

4. `FockSpace` is just a tuple of tuples of occupation numbers. However, for
most calculations, it is easier and faster to use `np.ndarray`, especially if
the calculation can be vectorized.

TBD