fix(setup.py): Blackbird + new Numpy fails
`quantum-blackbird==0.5.0` raises the following error with `numpy==1.24.3˙:

```
AttributeError: module 'numpy' has no attribute 'str'.
`np.str` was a deprecated alias for the builtin `str`. To avoid this error in existing code, use `str` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.str_` here.
The aliases was originally deprecated in NumPy 1.20; for more details and guidance see the original release note at:
    https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
```

In the main branch of Blackbird, this issue has already been solved, but it has
not yet been published. A temporary solution is to include the URL of the
repository in `setup.py`, which has been done in this patch.