fix(tensorflow): Executing with initial state
**Problem**

When an initial state is specified for `Simulator.execute` and `tf.function` is
used, Piquasso throws an error, because the specified initial state is
deep-copied, and TensorFlow variables cannot be copied like this.

**Solution**

A `PureFockState.copy` method has been written, which copies explicitely the
state vector using the NumPy API of TensorFlow.