perf(hafnian): Replace implementation
**Problem**

The PiquassoBoost version of the (loop)hafnian is very efficient, but the
implementation in Piquasso itself is not, the performance gap is large between
the two implementations.

**Solution**

An efficient version of the (loop)hafnian function got implemented using Numba
and the PiquassoBoost code translated from C++ to Python. The power trace
algorithm has been copied using Glynn-type iterations, which is enabled by
default in PiquassoBoost. Of course, this does not replace the PiquassoBoost
implementation, only narrows the performance gap between the two, to provide
relatively fast (loop)hafnian calculations without the need of installing
PiquassoBoost.

`numba` is added as a dependency.

The `piquasso/_math/hafnian` folder is ignored in Codecov, since it could not track code compiled with Numba.