feat!(config): Delete `normalize` variable
The normalization in calculations are redundant, because the normalization is
just a multiplier, and may be moved after all the gates are executed. However,
this normalization may nontrivially affect measurement results, hence a
`pq.Normalize()` gate is created, which may or may not be used before
measurement as required. The `(Pure)FockState.normalize()` method is kept, in
case the user wishes to normalize the state.

Moreover, the test `test_PureFockState_displaced_state_variance_photon_number`
has been fixed, since it was incorrectly parametrized.