feat(sampling): Custom permanent
**Problem**

In order to override the permanent calculation in `SamplingSimulator`, one had
to override a calculation function in `SamplingSimulator._instruction_map`,
which is what PiquassoBoost did. However, it was not possible to directly
specify the permanent. A permanent function could be provided in
`BaseCalculator`, but that was only used by `(Pure)FockSimulator` until
66bc621933532fbd3c80f377c123f10775443e57.

**Solution**

A permanent function has been implemented according to the algorithm described
in [High performance Boson Sampling simulation via data-flow
engines](https://arxiv.org/pdf/2309.07027.pdf), which has been translated from
PiquassoBoost to Python and Numba.

The problem with `theboss` is that it is tricky to specify variables from
`Config` (e.g., `dtype`, `seed_sequence`) or `BaseCalculator` classes. It would
be easier to handle these issues if we would not explicitely depend on this
contribution. Therefore, dependency on `theboss` has been dropped, and the
relevant code has been copied and rewritten.