fix(fock): Making `fock_probabilities` differentiable
The `fock_probabilities`, and hence `norm` and `normalize` were not
differentiable in `PureFockState`. This has been rewritten to use
`calculator.np`, but it broke `test_ImperfectPostSelectPhotons_gradient`.
Apparently, the tensorflow varible has been cast to a numpy variable in
`fock_probabilities`, and the calculation of the gradient did not factor in the
normalization in this test.