feat(sampling): Partial JAX support
A partial JAX support in `SamplingSimulator` has been enabled. For this, the
permanent is implemented in JAX. The same algorithm could not be implemented as
in `piquasso._math.permanent`, since JIT compilation in XLA does not handle
dynamic shapes which depend on non-static values, e.g., row or column
repetitions. Therefore, the permanent is needed to be reduced before
calculating the permanent with the Glynn algorithm using Gray codes is
implemented.

Moreover, the method `assym_reduce` has been rewritten, which produces the
reduced (or scattering) matrix.

The gradient of some attributes of `SamplingState` also works, and is tested in
`sampling/test_jax_gradient.py` and `test_backend_gradient_equivalence.py`.