feat!(sampling): Support for combining input states
So far, it has only been possible to specify a single occupation number in
`SamplingSimulator`.

In this patch, support for specifying multiple occupation numbers has been
specified.

The breaking change is that the length of the returned state vector and
probability distribution now depends on `Config.cutoff` instead of the
particles in the provided occupation numbers.