feat(combinatorics): Partitions Numba JIT
The `partitions` function now can be JIT compiled using Numba. For this
solution the `nb_combinations` function had to be defined which is
equivalent to `itertools.combinations` but numba compatible. The import
remained since other functions still use the `itertools` implementation
as well.

Note: The shape `(1, 0)` is necessary so that all return paths have the
same dimensionality. Furthermore, it is spicifically `1` in the first
dimension so that later on in `get_fock_space_basis` we will be able to
stack them up correctly when counting the elements for each item in the
lists.