perf(fock): jitted `calculate_interferometer_helper_indices` (#321)
Several functions have been rewritten so that they can be compiled using
Numba JIT. This is a partial milestone and more functions are planned to
be rewritten in this manner.

Specifically, the following changes occurred:
- The `comb` function is now part of the `combinatorics.py` module, and
an additional `arr_comb` is implemented accepting and returning arrays.
- `partitions` is now also jitted alongside `nb_combinations`,
equivalent to `itertools.combinations` which is used in many places,
hence the different name with the `nb_` prefix.
- The function `get_fock_space_basis` has also a new implementation with
a `nb_` prefix denoting the jitted version. The distinction is necessary
due to the extensive use of LRU caching at other parts of the code.
- The function `get_index_in_fock_subspace_array` now can be compiled
using Numba JIT.
- And finally `calculate_interferometer_helper_indices` is also now
jitted with Numba.

Notes:
- For the special case of the parametrization of `partitions` the shape
`(1,0)` is specifically chosen so that it can be stacked in
`get_fock_space_basis` correctly.
- The function header for `cutoff_cardinality_array` got changed so that
it does not force keyword arguments. This seems to be a bug in Numba
related to keyword arguments when they are called in another function
that is also JIT compiled. (see issue
[here](https://github.com/numba/numba/issues/9185))
- The function `calculate_interferometer_helper_indices` now returns
tuples instead of dictionaries (due to Numba) and therefore a temporary
dictionary is added between function calls until the entire
`get_interferometer_on_fock_space` function gets Numba JIT
implementation as well.

---------

Co-authored-by: Zoltán Kolarovszki <ifj.kolarovszki@gmail.com>