perf(fock): Interferometer on Fock space JITted
Calculating the unitary matrix corresponding to an interferometer is rewritten
to be `numba.njit`-compatible. For this, the `helper_indices` dict is rewritten
to be a tuple for compatibility.

The Numba implementation has been drastically modified compared to the Numpy
implementation to increase performance. This is because using Numpy one wants
to avoid, e.g., Python for loops, whereas in the JIT-compiled Numba code it is
more efficient to use explicit for loops.