perf(torontonian): Replace torontonian using C++
The previous torontonian implementation was slow. In order to make this more
efficient, we need to use a different implementation. For this, we took and
adapted the PiquassoBoost implementation, which is written in C++.

In order to build C++ code, we use `pybind11` + `cmake`. For building and
repairing wheels, we use `cibuildwheel`.

Uploading to PyPI will be implemented in a different patch. Until that, one can
manually upload packages if needed.

Source:
- https://arxiv.org/abs/2109.04528