chore: Remove strawberryfields from benchmark dependencies
Strawberry Fields has not been updated for a while, and it fails with recent
version of SciPy (e.g., with strawberryfields==0.23.0 and scipy==1.14.0). For
this reason, the strawberryfields imports from `benchmarks` were moved to a
pytest fixture. In `scripts` this has not been done, since these files don't
use pytest, and have less rationale in running these without strawberryfields.