feat(fermionic): Fermionic Gaussian simulator
Experimental `fermionic` package with Gaussian simulations.

The package has been put under `piquasso.fermionic`, and its tests in
`tests/fermionic`. The reason is that I did not want to mix it with the
photonic simulators.

Also, some photonic instructions (e.g., `Beamsplitter`, `Interferometer`) are
reused for the fermionic simulations. However, the docstrings still refer to
photonic gates, this will need to be resolved in the future.