feat!(api): Make `BaseConnector` abstract
`BaseConnector` was not abstact before, one could technically instantiate it,
but basically all methods raise a `NotImplementedError`. This is wrong, since
in theory if there is a plugin writing a custom `BaseConnector`, simulations
could fail between minor versions, since we can't guarantee that we won't start
using a connector method in a calculation not used before.