perf(sampling): Enhance classical BS simulation
The classical simulation of Boson Sampling is implemented with a rudimentary
algorithm, but it got reimplemented using the Clifford&Clifford algorithm B,
see https://arxiv.org/abs/1706.01260.

Moreover, the implementation for the repeated BB/FG formula with n-ary Gray
code has been copied from PiquassoBoost.

Additional tests were written for the permanent, and additional scripts were
written for both the permanent and for BS as well. In
`boson_sampling_benchmark.py`, Piquasso is compared with Perceval in BS.

Some seeded tests in `tests/_simulators/sampling` had to be modified  due to
the modification of the sampling algorithm.