perf(gaussian): Improve GBS simulation performance
Although a loop hafnian implementation has been moved from PiquassoBoost which
is significantly fast, Piquasso still lacks batching of the loop hafnian, as
described in Appendix B.5 of https://arxiv.org/abs/2108.01622, which would
yield a performance increase.

Therefore, this strategy has been implemented. One should note, that the
Glynn-type iterations could also be halved in the case of the batch hafnian,
and in the batch loop hafnian, it can be arranged that the even and odd
occupation number cases have different independent loops that can be halved.
The batch hafnian is not used in production code, but is kept as a reference
(this strategy could also be used for density matrix calculation later).

The sampling algorithm itself got also updated and made slightly more
efficient. In order to verify that the algorithm is correct, several scripts
have been rewritten, and a hypothesis testing has been written in `tests/slow`.
More specifically, I deleted the old `cramer_hypothesis_test` fixture because
it depends on an R library, and use a simple chi-squared hypothesis testing.
`ThresholdMeasurement` is also tested, since it uses the same algorithm by
default (`use_torontonian=False`), but it is also tested with
`use_torontonian=True`, because the tests only differ by a single line.

A benchmark script has been written in
`scripts/gaussian_boson_sampling_benchmark.py`.