chore: Add Python 3.13 support
Python 3.13 support is added in this patch.

It was a bit tricky, as TensorFlow is not released for Python 3.13 yet, so one
has to disable all the tests that use TensorFlow on Python 3.13. This is done
by creating a new `--skip-tensorflow` pytest option, which disables these
tests. By putting the `tensorflow` import in a fixture, we can optionally skip
those tests that try to use this fixture.

As there are many parametrized tests depending on `tensorflow` (e.g., through
`TensorflowConnector`), we had to implement a mechanism to use fixtures in
parametrized tests as parameters. This is enabled by the `pytest_lazy_fixtures`
plugin.

The test `tests/_simulators/fock/pure/test_state.py::test_mean_position`
mistakenly used `TensorflowConnector`, so this test has been rewritten to use
the default `NumpyConnector` instead.

The `--skip-tensorflow` flag is specified in `tox.ini` conditionally.

Source:
- [Pytest lazy fixtures plugin](https://github.com/dev-petrov/pytest-lazy-fixtures)
- [tox conditional settings](https://tox.wiki/en/latest/config.html#conditional-settings)