fix(fermionic): Majorana string sorting fix
The sorting of the Majorana string was done with `np.argsort`, but the
required permutation might be ambiguous when multiplicities are present,
messing up the parity term. When commuting Majorana operators, one would
naively do a bubble sort, which avoids commuting equal elements that would
introduce an unwanted extra term. Faster sorting algorithms might not adhere to
this, and the resulting permutation might not reflect the parity of the string.
In `numpy<2.0.0`, the resulting permutation of the sorting algorithm had a
proper parity, but in `numpy==2.0.2`, the algorithm is changed and results in
the wrong parity. Therefore, the sorting is reimplemented with a bubble sort,
which also calculates the parity term.