refact(_math): xpxp-xxpp basis changing rewritten
Sometimes we need to switch between the xxpp and xpxp basis. In Piquasso, this
is implemented by a basis changing matrix. However, as this matrix only
reorders rows/columns, it would be much nicer if the vectors/matrices would
just be reindexed. Therefore, this approach is deleted in favour of basis
changing indices.