feat(fermionic): `PureFockState.fock_probabilities_map` (#445)
In the `fermonic` package, the method `PureFockState.fock_probabilities`
returns with the particle number detection probabilities for a generic
pure state written in the Fock basis. However, the corresponding
occupation numbers are not returned, but the ordering explained in the
[documentation](https://piquasso.readthedocs.io/en/v6.0.1/experimental/fermionic.html#piquasso.fermionic.gaussian.GaussianState.fock_probabilities).
This might be inconvenient for users.

To remedy this, the `PureFockState.fock_probabilities_map` method
(practically a property, as `fock_probabilities`) is implemented, which
returns `fock_probabilities` "zipped" with the corresponding occupation
numbers.

Closes #435.