Suggestions
I rewrote the tests to test against analytical formulas obtained from
conjugating quadrature operators, and extended the docstring with a
justification for the implemented formula. I also renamed some variables, and
renamed the method itself to `get_xp_string_moment` instead, because XP-strings
are not necessarily observables.

I also deleted the implementation in `BaseFockState`, because
- I think the tests for `GaussianState` are pretty thorough now,
- It can yield very different results compared to the analytic formula,
- It also is quite slow.
I have a better implementation for this in mind, but I think it is not in scope
for this PR.